+++ cli/tests/init_compile.sh
@@ -1,4 +1,8 @@
 # docker is needed for this
+if [ -d "output/init-compile" ]; then
+    rm -rf output/init-compile
+fi
+
 mkdir -p output/init-compile
 cd output/init-compile