+++ packages/core/src/http/types/docsConfiguration.types.ts
@@ -1,0 +1,9 @@
+import { ApiReferenceConfiguration } from '@scalar/express-api-reference';
+
+export type DocsConfiguration =
+  | ({
+      type: 'scalar';
+    } & Partial<Omit<ApiReferenceConfiguration, 'spec'>>)
+  | {
+      type: 'swagger';
+    };