+++ packages/core/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @forklaunch/core
 
+## 0.5.6
+
+### Patch Changes
+
+- Increase package dependency versions
+- Updated dependencies
+  - @forklaunch/validator@0.4.7
+  - @forklaunch/common@0.2.4
+
+## 0.5.5
+
+### Patch Changes
+
+- Enables docs configuration to be set by caller and sends parsing error information to client if api parsing fails
+
 ## 0.5.4
 
 ### Patch Changes