+++ packages/express/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @forklaunch/express
 
+## 0.3.7
+
+### Patch Changes
+
+- Increase package dependency versions
+- Updated dependencies
+  - @forklaunch/validator@0.4.7
+  - @forklaunch/common@0.2.4
+  - @forklaunch/core@0.5.6
+
+## 0.3.6
+
+### Patch Changes
+
+- Enables docs configuration to be set by caller and sends parsing error information to client if api parsing fails
+- Updated dependencies
+  - @forklaunch/core@0.5.5
+
 ## 0.3.5
 
 ### Patch Changes